pyproject.toml
[tool.poetry]
name = "osintkit"
version = "0.1.0"
description = "OSINT CLI tool for personal digital footprint analysis"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "osintkit"}]

[tool.poetry.dependencies]
python = ">=3.10"
typer = {extras = ["all"], version = ">=0.12.0"}
httpx = "^0.27.0"
rich = "^13.7.0"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"
pydantic = "^2.5.0"
aiofiles = "^23.2.1"
phonenumbers = ">=8.13"
questionary = ">=2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-httpx = "^0.30"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
osintkit = "osintkit.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "live: marks tests requiring real network",
]
